hotelbase/backend/pyproject.toml
[project]
name = "hotel-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "mailersend>=2.0.0",
    "openpyxl>=3.1.5",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
]
